Fix logic to always select the best account, not the "next-best". i.e. if you had passwordless, github and LDAP, and you logged in as passwordless, ie would select github first and tell you to login as github. Then you login as github, and it would select LDAP and tell you to login as LDAP. Then finally, you login as LDAP and it works. Terrible UX :)
This code will now directly select LDAP in this case and tell you to
login as LDAP, regardless of you trying to login as passwordless or
GitHub first.

Additionally, this now sends the `redirect_uri` with the error message,
as a parameter to the SSO Dashboard, which may make use of this in the
future to send the user back.