Change access expiration logic to occur after AccessRules have succeeded This involves some refactoring.
Note that this also means:
- accessing *any* RP will update your authoritativeGroups.lastUsed
timestamp for that RP
- only RPs with an expiration setting in apps.yml will enforce the
access expiration at this time. We may enable it by default everywhere
in the future.
- if you are granted access to an RP, but never use the access, your
timestamp in lastUsed is never set. This means you can keep that access
and it will NOT expire until you first login to the RP, and after the
lastUsed attribute expires. This is a known issue that we're accepting
as expiration of access is yet another additional level of security, not the
primary decision for access.

We may be able to fix the issue eventually, if groups get tied to roles,
which themselves must be tied to an RP-subject for example (so that we
know why RPs you have access to, and thus set your lastUsed timestamp as
soon as you get access, regardless of you logging in or not)