Set a new claim https://sso.mozilla.com/claim/original_connection_user_id
This passes the user_id that the user would normally log in as if there
were no account linking.

Eg:
Case 1:
unlinked GitHub account logins, the user_id is: github|628537
original original_connection_user_id will also be github|628537.

Case 2:
Main/primary FxA account has a linked GitHub account, and the user logs
in with GitHub. The user_id is: oauth2|firefoxaccounts|3133731337
The original_connection_user_id is github|628537 in that case.

This code is mainly to be used by the Mozillians.org verification client
as it needs to know which IdP was used to perform authentication