Change sleep length to length of TOTP validity when invalid TOTP Duo code encountered
When parallelized test cause the same test user to do more than one Duo TOTP
authentication in the same 30 second window during which the TOTP doesn't change
Duo rejects the code as a replay. This increases the sleep in this case to 30
seconds to ensure a new TOTP code is generated.