Change ldap autologin test to run after the non-autologin test
This collapses the two ldap login tests into one so we can test
both the login and autologin using the same session.

This also changes the autologin from testing production to testing
dev (previously the ldap autologin test tested sso.mozilla.com and
discourse which are production sites). This now autologins to the
base_url