Change negative scope testing to positive
Previously we were looking if the scopes requested were only email, openid or both. In that case  we would skip the fixup.
This now instead looks to see if there are any scopes of either profile or that begin with https://. If there aren't any of those scopes, then fixup is skipped
This also no longer uses lodash as we don't need it.