Change how BitSight role mapping works
Turns out that these roles are used with each log in and will change an existing user's role (as opposed to only being using on new user provisioning). As a result, we need these roles to map to user groups.

This commit was rewritten on 2021-09-16 in order to repair a missing GPG
signature. I authorized these changes previously and I continue to
endorse them at the time of this author rewrite and gpg signature.
Prior to the rewrite, this commit was:

0fa187fa7bd7b4d5a24eb59da56bab1721208052
From: Gene Wood <gene_wood@cementhorizon.com>
Date: Mon, 2 Aug 2021 16:26:02 -0700

-- 20210916 atoll