ghe Deny because security has alerting set up
This commit removes the redirect functionality we used to have, in favour of
plain denying the authentication request.

Security ingests these events and needs to monitor them. And by redirecting
users to the wiki page, we short-circuit their monitoring pipeline.

In lieu of a redirect to the wiki page when experiencing issues logging in,
users will need to reach out to the GHE admins for help (which they would have
had to do anyways).

Should the frequency of login issues increase, Security said they'd be open to
creating a shared Security+GHE Admins channel, where Auth0 audit events are
sent, which would include the error message we send with the wiki page. This
means that the GHE Admins would, at the very least, be able to see the reason
why our auth pipeline rejected the login attempt.

Jira: IAM-1613